generate all http headers required
The `sign!` method in
[api_auth](https://github.com/mgomes/api_auth/blob/4ea2e2f6abed1fdf650ec716ca86c2ed9f640b45/spec/api_auth_spec.rb#L26)
adds more than just the `Authorization` header. It also adds the
following headers:

* `DATE`
* `Content-MD5`

This commit adds those headers in a way that should be compatible with
`api_auth`. It replaces the `authorization` method with `headers`, and
instead of returning a string, returns a map of all of the headers that
should be included in the request in order to be authenticated by
`api_auth`.

It also makes a few other small changes to match the behavior of
`api_auth`