README Documentation for RubyGem Compatibility
This updates the README's compatibility section to add a bit more
context, in the way of headings, for scanning which section of the
documentation is relevant, depending on your use case.

This also updates that section's code snippets to specify variables
which can and should be atoms that are argument's passed to the
`authentic?` or `headers` functions.